In this file I've started to creat the necessary functions to handle setting game_state. I've noticed we lack clarity of how we are going to stablish boats setting by user.
In particular, we have to decide whether user selects first the length of the boat and then clicks on cells to locate it or they click the cells straight ahead and then we try a match of corresponding length of selection with the boats available (this last option is the one that I was preparing). Either way, I insist that we inform the user of what error they've comitted when trying to set boat.

Also, Sergio has suggested that when selecting a boat and a cell (first option up), we paint in advance the legal cells in which the boat can be set to help the user make an idea of the appearence.

I think that is going to make the project much more difficult, but we'll discuss it soon.